backend/requirements.txt
# FastAPI et dépendances
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy>=2.0.35
alembic>=1.14.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# API clients
httpx==0.26.0
pylast==5.2.0
python3-discogs-client==2.6
spotipy==2.23.0

# Background tasks
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
